Move Serializers into the Views
