+++ apps/rabbitci_core/lib/ja_serializer_impls.ex
@@ -1,0 +1,9 @@
+defimpl JaSerializer.Formatter, for: [Task] do
+  alias JaSerializer.Formatter
+
+  def format(task) do
+    task
+    |> Task.await
+    |> Formatter.format
+  end
+end
