+++ apps/rabbitci_core/web/channels/branch_updater_channel.ex
@@ -3,7 +3,7 @@ defmodule RabbitCICore.BranchUpdaterChannel do
   alias RabbitCICore.Build
   alias RabbitCICore.Endpoint
 
-  def join("branches:" <> branch_id, payload, socket) do
+  def join("branches:" <> _branch_id, payload, socket) do
     if authorized?(payload) do
       {:ok, socket}
     else
@@ -17,8 +17,8 @@ defmodule RabbitCICore.BranchUpdaterChannel do
   end
 
   def new_build(branch_id, build_id) do
+    payload = Build.json_from_id!(build_id)
     Task.start fn ->
-      payload = Build.json_from_id!(build_id)
       Endpoint.broadcast("branches:#{branch_id}", "new:build", payload)
     end
   end
