+++ apps/rabbitci_core/web/views/build_view.ex
@@ -1,14 +1,23 @@
 defmodule RabbitCICore.BuildView do
   use RabbitCICore.Web, :view
-  alias RabbitCICore.BuildSerializer
+  use JaSerializer.PhoenixView
 
-  def render("index.json", %{conn: conn, builds: builds}) do
-    BuildSerializer.format(builds, conn, %{})
+  def render("running_builds.json", %{conn: conn, data: builds}) do
+    format(builds, conn, %{})
   end
-  def render("index.json", %{conn: conn, build: build}) do
-    BuildSerializer.format(build, conn, %{})
-  end
-  def render("running_builds.json", %{conn: conn, builds: builds}) do
-    BuildSerializer.format(builds, conn, %{})
-  end
+
+  alias RabbitCICore.Repo
+  alias RabbitCICore.Build
+  alias RabbitCICore.BranchView
+  alias RabbitCICore.StepView
+
+  attributes [:build_number, :start_time, :finish_time, :updated_at,
+              :inserted_at, :commit, :status, :config_extracted]
+  has_one :branch, include: true, serializer: BranchView
+  has_many :steps, include: true, serializer: StepView
+
+  def type, do: "builds"
+  def branch(r, _), do: Repo.preload(r, :branch).branch
+  def steps(r, _), do: Repo.preload(r, :steps).steps
+  def status(record), do: Build.status(record)
 end
